requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
pdf2image
pillow
python-dotenv
streamlit  
requests   